requirements.txt
pytorch_lightning==1.3.2
transformers==4.6.0
Pillow==8.1.0
tqdm==4.56.0
ipdb==0.13.4
numpy==1.19.5
einops==0.3.0
pyarrow==2.0.0
sacred==0.8.2
pandas==1.1.5
git+https://github.com/rwightman/pytorch-image-models.git
ftfy
torchvision~=0.8.2
torch~=1.7.1
